# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi documentation for Tree Nodes",
        "description": "Swagger OpenApi documentation for Tree Nodes",
        "contact": {
            "email": "info@website.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/nodes": {
            "get": {
                "tags": [
                    "nodes"
                ],
                "summary": "All nodes paginated",
                "description": "Display a listing of the resource.\nShowing all nodes paginated.",
                "operationId": "6a49c5145f15e738bcec4f33191ed0c4",
                "parameters": [
                    {
                        "name": "X-Header-Timezone",
                        "in": "header",
                        "description": "Get date fields in specific Timezone",
                        "required": false,
                        "examples": {
                            "Caracas": {
                                "summary": "Caracas",
                                "value": "America/Caracas"
                            },
                            "Bogota": {
                                "summary": "Bogota",
                                "value": "America/Bogota"
                            },
                            "Los Angeles": {
                                "summary": "Los Angeles",
                                "value": "America/Los_Angeles"
                            }
                        }
                    },
                    {
                        "name": "X-Header-Lang",
                        "in": "header",
                        "description": "Get title field in specific language",
                        "required": false,
                        "examples": {
                            "en": {
                                "summary": "English",
                                "value": "en"
                            },
                            "es": {
                                "summary": "Spanish",
                                "value": "es"
                            },
                            "fr": {
                                "summary": "French",
                                "value": "fr"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All nodes paginated."
                    },
                    "500": {
                        "description": "An error has happened."
                    }
                }
            },
            "post": {
                "tags": [
                    "nodes"
                ],
                "summary": "Storing a new node",
                "description": "Store a new created node in storage.\nStoring a new node in a tree or creating root node for a new tree.",
                "operationId": "4211f0fc512d4ae4547f7ead6fbbfed3",
                "parameters": [
                    {
                        "name": "X-Header-Timezone",
                        "in": "header",
                        "description": "Get date fields in specific Timezone",
                        "required": false,
                        "examples": {
                            "Caracas": {
                                "summary": "Caracas",
                                "value": "America/Caracas"
                            },
                            "Bogota": {
                                "summary": "Bogota",
                                "value": "America/Bogota"
                            },
                            "Los Angeles": {
                                "summary": "Los Angeles",
                                "value": "America/Los_Angeles"
                            }
                        }
                    },
                    {
                        "name": "X-Header-Lang",
                        "in": "header",
                        "description": "Get title field in specific language",
                        "required": false,
                        "examples": {
                            "en": {
                                "summary": "English",
                                "value": "en"
                            },
                            "es": {
                                "summary": "Spanish",
                                "value": "es"
                            },
                            "fr": {
                                "summary": "French",
                                "value": "fr"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "parent": {
                                        "oneOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "integer"
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "parent": 1
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "New node created."
                    },
                    "404": {
                        "description": "Parent node not found."
                    },
                    "500": {
                        "description": "An error has happened."
                    }
                }
            }
        },
        "/api/nodes_with_children": {
            "get": {
                "tags": [
                    "nodes"
                ],
                "summary": "All nodes with children paginated",
                "description": "Display a listing of the resource.\nShowing all nodes with children paginated.",
                "operationId": "71c76059d699bc3ae4a14ecf9f0de74a",
                "parameters": [
                    {
                        "name": "X-Header-Timezone",
                        "in": "header",
                        "description": "Get date fields in specific Timezone",
                        "required": false,
                        "examples": {
                            "Caracas": {
                                "summary": "Caracas",
                                "value": "America/Caracas"
                            },
                            "Bogota": {
                                "summary": "Bogota",
                                "value": "America/Bogota"
                            },
                            "Los Angeles": {
                                "summary": "Los Angeles",
                                "value": "America/Los_Angeles"
                            }
                        }
                    },
                    {
                        "name": "X-Header-Lang",
                        "in": "header",
                        "description": "Get title field in specific language",
                        "required": false,
                        "examples": {
                            "en": {
                                "summary": "English",
                                "value": "en"
                            },
                            "es": {
                                "summary": "Spanish",
                                "value": "es"
                            },
                            "fr": {
                                "summary": "French",
                                "value": "fr"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All nodes with children paginated."
                    },
                    "500": {
                        "description": "An error has happened."
                    }
                }
            }
        },
        "/api/nodes_without_children": {
            "get": {
                "tags": [
                    "nodes"
                ],
                "summary": "All nodes without children paginated",
                "description": "Display a listing of the resource.\nShowing all nodes without children paginated.",
                "operationId": "b207385cec47e1113c9515e57dd489db",
                "parameters": [
                    {
                        "name": "X-Header-Timezone",
                        "in": "header",
                        "description": "Get date fields in specific Timezone",
                        "required": false,
                        "examples": {
                            "Caracas": {
                                "summary": "Caracas",
                                "value": "America/Caracas"
                            },
                            "Bogota": {
                                "summary": "Bogota",
                                "value": "America/Bogota"
                            },
                            "Los Angeles": {
                                "summary": "Los Angeles",
                                "value": "America/Los_Angeles"
                            }
                        }
                    },
                    {
                        "name": "X-Header-Lang",
                        "in": "header",
                        "description": "Get title field in specific language",
                        "required": false,
                        "examples": {
                            "en": {
                                "summary": "English",
                                "value": "en"
                            },
                            "es": {
                                "summary": "Spanish",
                                "value": "es"
                            },
                            "fr": {
                                "summary": "French",
                                "value": "fr"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All nodes without children paginated."
                    },
                    "500": {
                        "description": "An error has happened."
                    }
                }
            }
        },
...
    }
}